Added Filter (jQuery) menu and updated link label
This is clearly a hack, but it ”works”. I worked on this for many hours as JavaScript is not my expertise at all so if it has some bugs, I would not be surprised.

I had to sacrifice the seamless dynamic loading by adding a forced re-loading and inject jQuery which I did not want as it would make the website slower, but people did request a filter feature.